refactor : 인터페이스 제거 및 패키지 이동
