+++ src/test/java/domain/ShuffledCardsGeneratorTest.java
@@ -1,8 +1,8 @@
-package domain.strategy;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.Deck;
+import domain.ShuffledCardsGenerator;
 import domain.card.Card;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
